requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
streamlit>=1.35.0
streamlit-autorefresh>=1.0.1
pydeck>=0.9.0
folium>=0.16.0
streamlit-folium>=0.20.0
git+https://github.com/openskynetwork/opensky-api.git#egg=opensky-api&subdirectory=python
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.11.0
